ghex_msg_cb_mt:
this is the approach not suitable for callback-based architecture.
In the benchmark we have to wait for completion of each stage of inflight communication
requests. In this approach this requires a barrier, plus atomic operations
to increase / decrease the number of ongoing requests. In MPI this is much
more scalable, because each thread knows and checks it's own comm requests.
Similar good performance should be achievable by the futures aproach, which
is based on per-thread requests.