Adjust the eager/rendezvous switchover point when using GNI provider
GNI permits adjusting the eager/rendevzous threshold, but it only
appears to be beneficial when using a single thread, so set
new default values of 0x4000 and 0x400 for many/single threads
respectively and allow an override via an ENV var.

Use internal memory registrartion method instead of udreg as
it appears to be more reliable

Set default thread to 1 when thread_safe not requested

add extra GNI debugging of params. Clean up some set/get
of GNI params/vars.

Quiet a spurious error that GNI occasionally produces
with an error flag, but a success code in the error.